+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Throw UnknownUserException if attempting to load a user that doesn't exist 
 * Refactor DoctrineUserRepository to provide better extension point for custom
   saving of entities (e.g. to add entities that need to be saved with the user).
 * Require 0.3 series of warden-core and update repository method names